+++ lib/counter_culture/counter.rb
@@ -70,10 +70,16 @@ module CounterCulture
                           "#{model.connection.quote_column_name(change_counter_column)}"
                         end
 
+        column_type = klass.type_for_attribute(change_counter_column).type
+
         # we don't use Rails' update_counters because we support changing the timestamp
         updates = []
         # this updates the actual counter
-        updates << "#{quoted_column} = COALESCE(#{quoted_column}, 0) #{operator} #{delta_magnitude}"
+        if column_type == :money
+          updates << "#{quoted_column} = COALESCE(CAST(#{quoted_column} as NUMERIC), 0) #{operator} #{delta_magnitude}"
+        else
+          updates << "#{quoted_column} = COALESCE(#{quoted_column}, 0) #{operator} #{delta_magnitude}"
+        end
         # and here we update the timestamp, if so desired
         if touch
           current_time = obj.send(:current_time_from_proper_timezone)
@@ -158,8 +164,9 @@ module CounterCulture
     #   pass true to get the past value, false or nothing to get the
     #   current value
     def foreign_key_value(obj, relation, was = false)
+      original_relation = relation
       relation = relation.is_a?(Enumerable) ? relation.dup : [relation]
-      first_relation = relation.first
+      
       if was
         first = relation.shift
         foreign_key_value = attribute_was(obj, relation_foreign_key(first))
@@ -175,7 +182,8 @@ module CounterCulture
       while !value.nil? && relation.size > 0
         value = value.send(relation.shift)
       end
-      return value.try(relation_primary_key(first_relation, source: obj, was: was).try(:to_sym))
+
+      return value.try(relation_primary_key(original_relation, source: obj, was: was).try(:to_sym))
     end
 
     # gets the reflect object on the given relation
@@ -277,6 +285,8 @@ module CounterCulture
       if reflect.options.key?(:polymorphic)
         raise "can't handle multiple keys with polymorphic associations" unless (relation.is_a?(Symbol) || relation.length == 1)
         raise "must specify source for polymorphic associations..." unless source
+
+        return reflect.options[:primary_key] if reflect.options.key?(:primary_key)
         return relation_klass(relation, source: source, was: was).try(:primary_key)
       end
       reflect.association_primary_key(klass)
@@ -312,7 +322,9 @@ module CounterCulture
     end
 
     def execute_now_or_after_commit(obj, &block)
-      if @execute_after_commit
+      execute_after_commit = @execute_after_commit.is_a?(Proc) ? @execute_after_commit.call : @execute_after_commit
+
+      if execute_after_commit
         obj.execute_after_commit(&block)
       else
         block.call