+++ spec/schema.rb
@@ -156,6 +156,7 @@ ActiveRecord::Schema.define(:version => 20120522160158) do
     t.integer  "fk_cat_id"
     t.integer  "posts_count",       :default => 0, :null => false
     t.integer  "posts_after_commit_count",       :default => 0, :null => false
+    t.integer  "posts_dynamic_commit_count",       :default => 0, :null => false
     t.datetime "created_at"
     t.datetime "updated_at"
   end
@@ -216,20 +217,25 @@ ActiveRecord::Schema.define(:version => 20120522160158) do
   create_table "poly_images", :force => true do |t|
     t.integer "imageable_id", :null => true
     t.string "imageable_type", :null => true
+    t.string "imageable_uid", :null => true
     t.string "url"
   end
 
   create_table "poly_employees", :force => true do |t|
     t.string "name"
+    t.string "uid"
     t.integer  "poly_images_count", :default => 0, :null => false
     t.integer  "poly_images_count_dup", :default => 0, :null => false
+    t.integer  "poly_images_from_uids_count", :default => 0, :null => false
     t.integer  "special_poly_images_count", :default => 0, :null => false
   end
 
   create_table "poly_products", :primary_key => 'pp_pk_id', :force => true do |t|
     t.string "brand_name"
+    t.string "uid"
     t.integer  "poly_images_count", :default => 0, :null => false
     t.integer  "poly_images_count_dup", :default => 0, :null => false
+    t.integer  "poly_images_from_uids_count", :default => 0, :null => false
     t.integer  "special_poly_images_count", :default => 0, :null => false
   end
 
@@ -267,11 +273,36 @@ ActiveRecord::Schema.define(:version => 20120522160158) do
   create_table :prefectures, :force => true do |t|
     t.string :name
     t.integer :big_cities_count, null: false, default: 0
+    t.integer :small_cities_count, null: false, default: 0
   end
 
   create_table :cities, :force => true do |t|
     t.string :name
     t.integer :prefecture_id, null: false
     t.integer :population, null: false
+  end
+
+  create_table "groups", :force => true do |t|
+    t.integer  "group_items_count", :default => 0, :null => false
+  end
+
+  create_table "sub_groups", primary_key: 'uuid', id: false, :force => true do |t|
+    t.string  "uuid", :null => false
+    t.integer  "group_id", :null => false
+  end
+
+  create_table "group_items", :force => true do |t|
+    t.string  "sub_group_uuid", :null => false
+  end
+
+  if ENV['DB'] == 'postgresql' && Gem::Version.new(Rails.version) >= Gem::Version.new('5.0')
+    create_table :purchase_orders, :force => true do |t|
+      t.money "total_amount", scale: 2, default: "0.0", null: false
+    end
+
+    create_table :purchase_order_items, :force => true do |t|
+      t.integer :purchase_order_id, null: false
+      t.money "amount", scale: 2, default: "0.0", null: false
+    end
   end
 end